refactor: Remove update-specific code from form wizard controller
