+++ app/move/controllers/update/base.test.js
@@ -1,3 +1,6 @@
+const FormController = require('hmpo-form-wizard').Controller
+const { cloneDeep } = require('lodash')
+
 const proxyquire = require('proxyquire')
 const personService = require('../../../../common/services/person')
 
@@ -79,7 +82,7 @@ describe('Move controllers', function() {
       })
 
       it('should call correct number of additional middleware', function() {
-        expect(controller.use).to.be.callCount(1)
+        expect(controller.use).to.be.callCount(2)
       })
     })
 
@@ -153,7 +156,6 @@ describe('Move controllers', function() {
           },
         },
       }
-      const res = {}
       const person = { id: '#personId' }
       const values = { foo: 'bar' }
 
@@ -163,12 +165,12 @@ describe('Move controllers', function() {
       })
 
       it('should return updated values', function() {
-        expect(controller.getUpdateValues(req, res)).to.equal(values)
+        expect(controller.getUpdateValues(req)).to.equal(values)
       })
 
       context('When a valid ', function() {
         this.beforeEach(function() {
-          controller.getUpdateValues(req, res)
+          controller.getUpdateValues(req)
         })
         it('should get person data', function() {
           expect(req.getPerson).to.be.calledOnceWithExactly()
@@ -178,6 +180,57 @@ describe('Move controllers', function() {
             'field1',
             'field2',
           ])
+        })
+      })
+    })
+
+    describe('#setInitialStep()', function() {
+      let req
+      let nextSpy
+      const res = {}
+      beforeEach(function() {
+        req = {
+          get: sinon.stub(),
+          form: {
+            options: {
+              backLink: '/baseUrl',
+            },
+          },
+        }
+        nextSpy = sinon.stub()
+      })
+
+      it('should invoke the next middleware', function() {
+        controller.setInitialStep(req, res, nextSpy)
+        expect(nextSpy).to.be.calledOnceWithExactly()
+      })
+
+      context('when there is no referrer', function() {
+        it('should set initialStep to true', function() {
+          controller.setInitialStep(req, res, nextSpy)
+          expect(req.initialStep).to.be.true
+        })
+      })
+
+      context('when referrer is the base page', function() {
+        beforeEach(function() {
+          req.get.withArgs('referrer').returns('http://example.com/baseUrl')
+          controller.setInitialStep(req, res, nextSpy)
+        })
+
+        it('should set initialStep to true', function() {
+          expect(req.initialStep).to.be.true
+        })
+      })
+
+      context('when the referrer is anything else', function() {
+        beforeEach(function() {
+          req.get.withArgs('referrer').returns('http://example.com/someUrl')
+          controller.setInitialStep(req, res, nextSpy)
+        })
+
+        it('should not set initialStep to true', function() {
+          expect(req.initialStep).to.not.be.true
         })
       })
     })
@@ -318,6 +371,216 @@ describe('Move controllers', function() {
           expect(controller.superMethodSource()).to.equal('BaseSource')
         })
       })
+    })
+  })
+
+  describe('#getErrors()', function() {
+    let getErrorsStub
+    let req
+    let res
+
+    beforeEach(function() {
+      getErrorsStub = sinon.stub(BaseProto, 'getErrors')
+      req = {}
+      res = {}
+    })
+
+    context('when req.initialStep is false', function() {
+      it('should call the parent getErrors method', function() {
+        controller.getErrors(req, res)
+        expect(getErrorsStub).to.be.calledOnceWithExactly(req, res)
+      })
+    })
+
+    context('when req.initialStep is true', function() {
+      beforeEach(function() {
+        req.initialStep = true
+      })
+
+      it('should not call the parent getErrors method', function() {
+        controller.getErrors(req, res)
+        expect(getErrorsStub).to.not.be.called
+      })
+
+      it('should return an empty object', function() {
+        expect(controller.getErrors(req, res)).to.deep.equal({})
+      })
+    })
+  })
+
+  describe('#getValues()', function() {
+    let getValuesStub
+    let callback
+    let req
+    let res
+
+    beforeEach(function() {
+      callback = sinon.spy()
+      sinon.stub(controller, 'getUpdateValues').returns({ baz: 'tastic' })
+      sinon.stub(controller, 'protectReadOnlyFields')
+      getValuesStub = sinon.stub(FormController.prototype, 'getValues')
+      getValuesStub.callsFake((req, res, valuesCallback) => {
+        valuesCallback(null, { foo: 'bar' })
+      })
+      req = {
+        form: {
+          options: {},
+        },
+      }
+      res = {
+        locals: {},
+      }
+    })
+
+    context('when req.initialStep is false', function() {
+      it('should not call the getUpdateValues method', function() {
+        controller.getValues(req, res, callback)
+        expect(controller.getUpdateValues).to.not.be.called
+      })
+
+      it('should call the protectReadOnlyFields method with the correct args', function() {
+        controller.getValues(req, res, callback)
+        expect(controller.protectReadOnlyFields).to.be.calledOnceWithExactly(
+          req,
+          {
+            foo: 'bar',
+          }
+        )
+      })
+
+      it('should invoke the callback', function() {
+        controller.getValues(req, res, callback)
+        expect(callback).to.be.calledOnceWithExactly(null, { foo: 'bar' })
+      })
+    })
+
+    context('when req.initialStep is true', function() {
+      beforeEach(function() {
+        req.initialStep = true
+      })
+      it('should call the getUpdateValues method with the correct args', function() {
+        controller.getValues(req, res, callback)
+        expect(controller.getUpdateValues).to.be.calledOnceWithExactly(req)
+      })
+      it('should call the protectReadOnlyFields method with the correct args', function() {
+        controller.getValues(req, res, callback)
+        expect(controller.protectReadOnlyFields).to.be.calledOnceWithExactly(
+          req,
+          {
+            baz: 'tastic',
+          }
+        )
+      })
+
+      it('should invoke the callback with the correct args', function() {
+        controller.getValues(req, res, callback)
+        expect(callback).to.be.calledOnceWithExactly(null, {
+          baz: 'tastic',
+        })
+      })
+    })
+
+    context('when super.getUpdateValues passes an error', function() {
+      let err
+      beforeEach(function() {
+        err = new Error()
+        getValuesStub.callsFake((req, res, valuesCallback) => {
+          valuesCallback(err, { foo: 'bar' })
+        })
+      })
+
+      it('should invoke the callback with the error', function() {
+        controller.getValues(req, res, callback)
+        expect(callback).to.be.calledOnceWithExactly(err)
+      })
+    })
+
+    context('when this.getUpdateValues throws an error', function() {
+      let err
+      beforeEach(function() {
+        req.initialStep = true
+        req.form.options.update = true
+        err = new Error()
+        controller.getUpdateValues.restore()
+        sinon.stub(controller, 'getUpdateValues').throws(err)
+      })
+
+      it('should invoke the callback with the error', function() {
+        controller.getValues(req, res, callback)
+        expect(callback).to.be.calledOnceWithExactly(err)
+      })
+    })
+  })
+
+  describe('#getUpdateValues()', function() {
+    const req = {}
+    beforeEach(function() {
+      req.getPerson = sinon.stub()
+    })
+    context('when default getUpdateValues method invoked', function() {
+      it('should return an empty object', function() {
+        const values = controller.getUpdateValues(req, {})
+        expect(values).to.deep.equal({})
+      })
+    })
+  })
+
+  describe('#protectReadOnlyFields()', function() {
+    const unprotectedComponent = {
+      component: 'a',
+      prop: false,
+      updateComponent: {
+        component: 'b',
+        prop: true,
+        anotherProp: true,
+      },
+    }
+    const protectedComponent = cloneDeep(unprotectedComponent)
+    protectedComponent.readOnly = true
+
+    const fields = {
+      unprotected: cloneDeep(unprotectedComponent),
+      protectedUndef: cloneDeep(protectedComponent),
+      protectedNull: cloneDeep(protectedComponent),
+      protected: cloneDeep(protectedComponent),
+    }
+    let req
+    const values = {
+      unprotected: 'unprotected-value',
+      protectedNull: null,
+      protected: 'protected-value',
+    }
+    beforeEach(function() {
+      req = {
+        form: {
+          options: {
+            fields: cloneDeep(fields),
+          },
+        },
+      }
+    })
+
+    it('should leave unprotected fields unchanged', function() {
+      controller.protectReadOnlyFields(req, values)
+      const { unprotected } = req.form.options.fields
+      expect(unprotected).to.deep.equal(fields.unprotected)
+    })
+    it('should leave protected fields that have no value unchanged', function() {
+      controller.protectReadOnlyFields(req, values)
+      const { protectedUndef } = req.form.options.fields
+      expect(protectedUndef).to.deep.equal(fields.protectedUndef)
+    })
+    it('should leave protected fields that have null value unchanged', function() {
+      controller.protectReadOnlyFields(req, values)
+      const { protectedNull } = req.form.options.fields
+      expect(protectedNull).to.deep.equal(fields.protectedNull)
+    })
+    it('should update component of protected fields that have a value', function() {
+      controller.protectReadOnlyFields(req, values)
+      const { protected: protectedField } = req.form.options.fields
+      expect(protectedField.component).to.equal('b')
+      expect(protectedField).to.have.property('prop', true)
+      expect(protectedField).to.have.property('anotherProp', true)
     })
   })
 })
